Fixed a massive bug in TTable. Engine looks to be working close to if not stronger than it was before from initial impression. TT bug solved many issues like poor scoring of game board. Will test against old engine to see if anything is still wrong.

+++ externs.h
@@ -129,8 +129,8 @@ extern int SquareDistance[SQ_ALL][SQ_ALL];
 
 
 //UCI input varibles for "go"
-extern int wtime; //time left on whites clock
-extern int btime; //black clock
+extern long wtime; //time left on whites clock
+extern long btime; //black clock
 extern int winc;
 extern int binc;
 extern int movestogo;
